move modal to it's own components and handle props
